+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+- Expanded Baubles Support
+- Added Baubles Support
+- Added ArrayUtil.toString method
+- Expanded MutableString
+
 ## 1.1.1 : 
 Fix mod not loading